+++ tests/test_0030_align.py
@@ -11,7 +11,7 @@ from shinestacker.algorithms.align_auto import AlignFramesAuto
 def test_align():
     try:
         img_1, img_2 = [read_img(f"examples/input/img-jpg/000{i}.jpg") for i in (2, 3)]
-        n_good_matches, M, img_warp = align_images(img_1, img_2)
+        n_good_matches, m, img_warp = align_images(img_1, img_2)
         assert img_warp is not None
         assert n_good_matches > 100
     except Exception:
@@ -21,7 +21,7 @@ def test_align():
 def test_align_homo():
     try:
         img_1, img_2 = [read_img(f"examples/input/img-jpg/000{i}.jpg") for i in (2, 3)]
-        n_good_matches, M, img_warp = align_images(
+        n_good_matches, m, img_warp = align_images(
             img_1, img_2, alignment_config={'transform': constants.ALIGN_HOMOGRAPHY})
         assert img_warp is not None
         assert n_good_matches > 10
@@ -32,19 +32,8 @@ def test_align_homo():
 def test_align_rescale():
     try:
         img_1, img_2 = [read_img(f"examples/input/img-jpg/000{i}.jpg") for i in (2, 3)]
-        n_good_matches, M, img_warp = align_images(
+        n_good_matches, m, img_warp = align_images(
             img_1, img_2, alignment_config={'subsample': 4})
-        assert img_warp is not None
-        assert n_good_matches > 10
-    except Exception:
-        assert False
-
-
-def test_align_ecc():
-    try:
-        img_1, img_2 = [read_img(f"examples/input/img-jpg/000{i}.jpg") for i in (2, 3)]
-        n_good_matches, M, img_warp = align_images(
-            img_1, img_2, alignment_config={'ecc_refinement': True})
         assert img_warp is not None
         assert n_good_matches > 10
     except Exception: